requirements.txt
asgiref==3.5.1
autopep8==1.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.4
idna==3.3
mailjet-rest==1.3.4
protobuf==3.19.4
pycodestyle==2.8.0
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
urllib3==1.26.9
pysqlite3==0.4.7
